restrict handover tab access
