+++ src/Frontend/Dfe.Complete/Pages/Projects/List/_AllProjectsLayout.cshtml
@@ -1,10 +1,18 @@
 ﻿@using Dfe.Complete.Pages.Projects.List
 @using Dfe.Complete.Constants
+@using MediatR
+@using Dfe.Complete.Utils;
+
+@inject ISender sender
+
 @model Dfe.Complete.Pages.Projects.List.AllProjectsModel
 @{
   Layout = "Shared/_Layout";
   var ariaCurrentValue = "page";
   string? IsCurrentNav(string tabName) => Model.CurrentNavigationItem == tabName ? ariaCurrentValue : null;
+
+  var userTeam = await User.GetUserTeam(sender);
+  bool UserHasTabAccess(string tabName) => UserTabAccessHelper.UserHasTabAccess(userTeam, tabName);
 }
 
 @section BeforeMainFullWidth
@@ -14,32 +22,45 @@
       <div class="moj-primary-navigation__nav">
         <nav class="moj-primary-navigation" aria-label="Primary navigation">
           <ul class="moj-primary-navigation__list">
+            @if (UserHasTabAccess(AllProjectsModel.HandoverNavigation))
+            {
+              <li class="moj-primary-navigation__item">
+                <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.HandoverNavigation)"
+                  href="/projects/all/handover">Handover</a>
+              </li>
+            }
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.HandoverNavigation)" href="/projects/all/handover">Handover</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.InProgressNavigation)"
+                href="/projects/all/in-progress/all">In progress</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.InProgressNavigation)" href="/projects/all/in-progress/all">In progress</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByMonthNavigation)"
+                href="/projects/all/by-month/conversions">By month</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByMonthNavigation)" href="/projects/all/by-month/conversions">By month</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByRegionNavigation)"
+                href="/projects/all/regions">By region</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByRegionNavigation)" href="/projects/all/regions">By region</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByUserNavigation)"
+                href="/projects/all/users">By user</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByUserNavigation)" href="/projects/all/users">By user</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByTrustNavigation)"
+                href="/projects/all/trusts">By trust</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByTrustNavigation)" href="/projects/all/trusts">By trust</a>
+              <a class="moj-primary-navigation__link"
+                aria-current="@IsCurrentNav(AllProjectsModel.ByLocalAuthorityNavigation)"
+                href="/projects/all/local-authorities">By local authority</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.ByLocalAuthorityNavigation)" href="/projects/all/local-authorities">By local authority</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.CompletedNavigation)"
+                href="@RouteConstants.CompletedProjects">Completed</a>
             </li>
             <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.CompletedNavigation)" href="@RouteConstants.CompletedProjects">Completed</a>
-            </li>
-            <li class="moj-primary-navigation__item">
-              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.StatisticsNavigation)" href="/projects/all/statistics">Statistics</a>
+              <a class="moj-primary-navigation__link" aria-current="@IsCurrentNav(AllProjectsModel.StatisticsNavigation)"
+                href="/projects/all/statistics">Statistics</a>
             </li>
           </ul>
         </nav>
