+++ src/Frontend/Dfe.Complete/Pages/Shared/_Header.cshtml
@@ -13,7 +13,7 @@
   string IsCurrentTab(string tabName) => Model?.CurrentTab == tabName ? currentTabStyling : "";
 
   var userTeam = await User.GetUserTeam(sender);
-  bool UserHasTabAccess(string tabName) => TabNavigationModel.UserHasTabAccess(userTeam, tabName);
+  bool UserHasTabAccess(string tabName) => UserTabAccessHelper.UserHasTabAccess(userTeam, tabName);
 }
 
 <header class="dfe-header" role="banner">
