+++ src/Tests/Dfe.Complete.Tests/Models/TabNavigationModelTests.cs
@@ -14,21 +14,4 @@ public class TabNavigationModelTests
 
         Assert.Equal("my-current-tab", tabNavigationModel.CurrentTab);
     }
-
-    [Fact]
-    public void UserHasTabAccess_UnprotectedTab_ReturnsTrue()
-    {
-        var result = TabNavigationModel.UserHasTabAccess(Arg.Any<ProjectTeam>(), "your-projects");
-        Assert.True(result);
-    }
-
-    [Theory]
-    [InlineData(ProjectTeam.BusinessSupport, false)]
-    [InlineData(ProjectTeam.RegionalCaseWorkerServices, true)]
-    [InlineData(ProjectTeam.London, true)]
-    public void UserHasTabAccess_ProtectedTab_ReturnsCorrectAccessForTeam(ProjectTeam userTeam, bool expectedPermission)
-    {
-        var result = TabNavigationModel.UserHasTabAccess(userTeam, "your-team-projects");
-        Assert.Equal(expectedPermission, result);
-    }
 }
